% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co.R
\docType{data}
\name{co}
\alias{co}
\title{Tibble con i dati di co per 8 stazioni della regione ABRUZZO}
\format{
Un tibble con 8 colonne e 5848 osservazioni
}
\usage{
co
}
\description{
Tibble con i dati di co per 8 stazioni della regione ABRUZZO
}
\keyword{datasets}
