% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm25.R
\docType{data}
\name{pm25}
\alias{pm25}
\title{Tibble con i dati di pm25 per 8 stazioni della regione ABRUZZO}
\format{
Un tibble con 8 colonne e 5848 osservazioni
}
\usage{
pm25
}
\description{
Tibble con i dati di pm25 per 8 stazioni della regione ABRUZZO
}
\keyword{datasets}
